refactor rename package query2 to query
