+++ src/test/java/com/codecool/SQLYourCSV/model/query/QueryParserTest.java
@@ -1,0 +1,5 @@
+package com.codecool.SQLYourCSV.model.query;
+
+class QueryParserTest {
+
+}
