--- src/test/java/com/codecool/SQLYourCSV/model/query2/QueryParserTest.java
@@ -1,9 +1,0 @@
-package com.codecool.SQLYourCSV.model.query2;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class QueryParserTest {
-
-}
